Loot ID is now int32

+++ Source/Galdrar/Loot.cpp
@@ -36,7 +36,7 @@ void ALoot::Activate(CharacterStats* ownerStats, CharacterStats* targetStats)
 
 FString ALoot::GetName(){ return name; }
 FString ALoot::GetDesc(){ return desc; }
-float ALoot::GetID() { return ID; }
+int32 ALoot::GetID() { return ID; }
 
 bool ALoot::IsStackable(){ return bStackable; }
 bool ALoot::IsActivatable(){ return bActivatable; }
